requirements.txt
torch
torchvision
transformers
Pillow
opencv-python
openai